+++ Paradigm/src/utopia/paradigm/shape/shape2d/Bounded.scala
@@ -1,7 +1,8 @@
 package utopia.paradigm.shape.shape2d
 
+import utopia.paradigm.enumeration.Alignment
 import utopia.paradigm.enumeration.Axis.{X, Y}
-import utopia.paradigm.enumeration.{Axis, Axis2D}
+import utopia.paradigm.shape.template.HasDimensions.HasDoubleDimensions
 import utopia.paradigm.shape.template.DoubleVectorLike
 
 /**
@@ -9,86 +10,153 @@ import utopia.paradigm.shape.template.DoubleVectorLike
   * @author Mikko Hilpinen
   * @since Genesis 15.5.2021, v2.5.1
   */
-// TODO: Rename to HasBounds
-trait Bounded extends Sized with Area2D
+trait Bounded[+Repr] extends HasBounds with Sized[Repr]
 {
 	// ABSTRACT --------------------------
 	
 	/**
-	  * @return Bounding box around this shape
+	  * Creates a copy of this item with modified bounds
+	  * @param newBounds Bounds for the new item
+	  * @return A copy of this item with the specified bounds
 	  */
-	def bounds: Bounds
-	
-	
-	// COMPUTED -------------------------
-	
-	/**
-	  * @return The top-left corner of this item's bounds (assuming positive size)
-	  */
-	def topLeft = bounds.position
-	/**
-	  * @return The top-right corner of this item's bounds (assuming positive size)
-	  */
-	def topRight = topLeft + X(width)
-	/**
-	  * @return The bottom-left corner of this item's bounds (assuming positive size)
-	  */
-	def bottomLeft = topLeft + Y(height)
-	/**
-	  * @return The bottom-right corner of this item's bounds (assuming positive size)
-	  */
-	def bottomRight = topLeft + size
-	/**
-	  * @return The leftmost x-coordinate of this item's bounds (assuming positive size)
-	  */
-	def leftX = topLeft.x
-	/**
-	  * @return The rightmost x-coordinate of this item's bounds (assuming positive size)
-	  */
-	def rightX = leftX + width
-	/**
-	  * @return The topmost y-coordinate of this item's bounds (assuming positive size)
-	  */
-	def topY = topLeft.y
-	/**
-	  * @return The bottom-most y-coordinate of this item's bounds (assuming positive size)
-	  */
-	def bottomY = topY + height
+	def withBounds(newBounds: Bounds): Repr
 	
 	
 	// IMPLEMENTED  ---------------------
 	
-	override def size = bounds.size
-	
-	override def contains[V <: DoubleVectorLike[V]](point: V) =
-		point.x >= leftX && point.y >= topY && point.x <= rightX && point.y <= bottomY
+	override def withSize(size: Size) = withBounds(Bounds(topLeft, size))
 	
 	
 	// OTHER    -------------------------
 	
 	/**
-	  * @param axis The targeted axis
-	  * @return The smallest coordinate within this item's bounds along the specified axis (assuming positive size)
+	  * @param f A mapping function for this item's bounds
+	  * @return A copy of this item with mapped bounds
 	  */
-	def minAlong(axis: Axis) = topLeft.along(axis)
-	/**
-	  * @param axis The targeted axis
-	  * @return The largesst coordinate within this item's bounds along the specified axis (assuming positive size)
-	  */
-	def maxAlong(axis: Axis) = topLeft.along(axis) + size.along(axis)
+	def mapBounds(f: Bounds => Bounds) = withBounds(f(bounds))
 	
 	/**
-	  * @param bounds A set of bounds
-	  * @return Whether this item's bounds overlap with the other set of bounds
+	  * Moves this item so that it's top-left coordinate is at the specified location
+	  * @param topLeft New top-left coordinate for this item
+	  * @param stretch Whether this item should be stretched so that the bottom-right corner is preserved (default = false)
+	  * @return A copy of this item where the top-left coordinate is at the specified location
 	  */
-	def overlapsWith(bounds: Bounds) =
-		Axis2D.values.forall { axis =>
-			maxAlong(axis) > bounds.minAlong(axis) && bounds.maxAlong(axis) > minAlong(axis)
+	def withTopLeft(topLeft: Point, stretch: Boolean = false) = {
+		if (stretch)
+			withBounds(Bounds(topLeft, size + (this.topLeft - topLeft)))
+		else
+			withBounds(Bounds(topLeft, size))
+	}
+	/**
+	  * Moves this item so that it's top-right coordinate is at the specified location
+	  * @param topRight New top-right coordinate for this item
+	  * @param stretch Whether this item should be stretched so that the bottom-left corner is preserved (default = false)
+	  * @return A copy of this item where the top-right coordinate is at the specified location
+	  */
+	def withTopRight(topRight: Point, stretch: Boolean = false) = {
+		if (stretch)
+			withBounds(Bounds(topLeft.withY(topRight.y), size + Vector2D(topRight.x - rightX, topY - topRight.y)))
+		else
+			withBounds(Bounds(topRight - X(width), size))
+	}
+	/**
+	  * Moves this item so that it's bottom-left coordinate is at the specified location
+	  * @param bottomLeft New bottom-left coordinate for this item
+	  * @param stretch Whether this item should be stretched so that the top-right corner is preserved (default = false)
+	  * @return A copy of this item where the bottom-left coordinate is at the specified location
+	  */
+	def withBottomLeft(bottomLeft: Point, stretch: Boolean = false) = {
+		if (stretch)
+			withBounds(Bounds(bottomLeft.withY(topY), size + Vector2D(leftX - bottomLeft.x, bottomLeft.y - bottomY)))
+		else
+			withBounds(Bounds(bottomLeft - Y(height), size))
+	}
+	/**
+	  * Moves this item so that it's bottom-right coordinate is at the specified location
+	  * @param bottomRight New bottom-right coordinate for this item
+	  * @param stretch Whether this item should be stretched so that the top-left corner is preserved (default = false)
+	  * @return A copy of this item where the bottom-right coordinate is at the specified location
+	  */
+	def withBottomRight(bottomRight: Point, stretch: Boolean = false) = {
+		if (stretch)
+			withBounds(Bounds(topLeft, size + (bottomRight - this.bottomRight)))
+		else
+			withBounds(Bounds(bottomRight - size, size))
+	}
+	
+	/**
+	  * Creates a copy of this item that has been positioned within the specified bounds using a specific alignment.
+	  * Doesn't modify the size of this item at all by default, even when it would not fit the specified set of bounds.
+	  * However, if 'cropToFit' is set to true, the size of this item may be reduced to fit the specified bounds.
+	  * This reduction doesn't preserve the shape of this item, however.
+	  *
+	  * If you wish to place this item completely within a set of bounds while also preserving shape, please use
+	  * fittedWithin(Bounds, Alignment) instead.
+	  * @param bounds A set of bounds
+	  * @param alignment Alignment to use when positioning this item within the specified bounds.
+	  *                  For example, if left alignment is used, this item will be located at the center-left position
+	  *                  of the specified bounds.
+	  * @return A relocated copy of this item
+	  */
+	def positionedWithin(bounds: Bounds, alignment: Alignment, cropToFit: Boolean = false) = {
+		if (cropToFit) {
+			val newSize = size.croppedToFitWithin(bounds.size)
+			withBounds(Bounds(alignment.position(newSize, bounds), newSize))
 		}
+		else
+			withTopLeft(alignment.position(size, bounds))
+	}
 	
 	/**
+	  * Creates a copy of this item that has been positioned and possibly scaled to fit within the specified bounds,
+	  * using the specified alignment. Preserves shape.
 	  * @param bounds A set of bounds
-	  * @return Whether this item lies completely within the specified set of bounds
+	  * @param alignment Alignment to use when positioning this item within the specified bounds.
+	  *                  For example, if left alignment is used, this item will be located at the center-left position
+	  *                  of the specified bounds.
+	  * @return A copy of this item that fits within the specified set of bounds
 	  */
-	def liesCompletelyWithin(bounds: Bounds) = bounds.contains(topLeft) && bounds.contains(topRight)
+	def fittedWithin(bounds: Bounds, alignment: Alignment) = {
+		val newSize = size.fittingWithin(bounds.size)
+		val newTopLeft = alignment.position(newSize, bounds)
+		withBounds(Bounds(newTopLeft, newSize))
+	}
+	
+	/**
+	  * Creates a copy of this item with scaled bounds (both size AND position). Preserves shape.
+	  * @param scaling A scaling modifier to apply to the bounds of this item
+	  * @return A scaled copy of this item
+	  */
+	def withScaledBounds(scaling: Double) = withBounds(bounds * scaling)
+	/**
+	  * Creates a copy of this item with scaled bounds (both size AND position). Doesn't preserve shape.
+	  * @param scaling A scaling modifier to apply to the bounds of this item (different for different axes)
+	  * @return A scaled copy of this item
+	  */
+	def withScaledBounds(scaling: HasDoubleDimensions) = withBounds(bounds * scaling)
+	
+	/**
+	  * @param translation Translation to apply to this item's position
+	  * @return A copy of this item with translated position
+	  */
+	def translated(translation: HasDoubleDimensions) = withBounds(bounds + translation)
+	
+	/**
+	  * @param insets Insets to apply to this item's bounds
+	  * @return A copy of this item with the insets added (both position and size are modified)
+	  */
+	def withInsets(insets: Insets) = withBounds(bounds + insets)
+	/**
+	  * Creates a copy of this item with enlarged bounds where the center-point remains the same
+	  * @param enlargement A size increase to apply
+	  * @return A copy of this item with bounds that keep the same center-point but have enlarged size
+	  */
+	def enlarged[V <: DoubleVectorLike[V]](enlargement: V) =
+		withBounds(Bounds(topLeft - enlargement / 2, size + enlargement))
+	/**
+	  * Creates a copy of this item with shrunk bounds where the center-point remains the same
+	  * @param shrinking A size decrease to apply
+	  * @return A copy of this item with bounds that keep the same center-point but have shrunk size
+	  */
+	def shrunk[V <: DoubleVectorLike[V]](shrinking: V) = withBounds(Bounds(topLeft + shrinking / 2, size - shrinking))
 }