+++ Paradigm/src/utopia/paradigm/shape/shape2d/Polygonic.scala
@@ -14,7 +14,7 @@ import java.awt.Shape
   * @since Genesis 14.4.2019, v2+
   */
 // TODO: Handle cases where there are 0 corners
-trait Polygonic extends ShapeConvertible with Projectable with Area2D with Transformable[Polygonic] with Bounded
+trait Polygonic extends ShapeConvertible with Projectable with Area2D with Transformable[Polygonic] with HasBounds
 {
 	// ABSTRACT	----------------