+++ Reflection/src/utopia/reflection/component/template/layout/Area.scala
@@ -3,7 +3,7 @@ package utopia.reflection.component.template.layout
 import utopia.paradigm.enumeration.Axis.{X, Y}
 import utopia.paradigm.enumeration.Axis2D
 import utopia.paradigm.shape.shape1d.Vector1D
-import utopia.paradigm.shape.shape2d.{Bounded, Bounds, Point, Size}
+import utopia.paradigm.shape.shape2d.{HasBounds, Bounds, Point, Size}
 import utopia.paradigm.shape.shape3d.Vector3D
 import utopia.paradigm.shape.template.HasDimensions.HasDoubleDimensions
 
@@ -12,7 +12,7 @@ import utopia.paradigm.shape.template.HasDimensions.HasDoubleDimensions
 * @author Mikko Hilpinen
 * @since 26.2.2019
 **/
-trait Area extends Bounded
+trait Area extends HasBounds
 {
     // ABSTRACT    ------------------------